requirements.txt
django
langchain
google-generativeai
python-dotenv
faiss-cpu
langchain_google_genai
langchain_community
PyMuPDF
django-admin-corporate==1.0.1
# pdf
python-docx
underthesea
mysqlclient
# openai
# tiktoken
# .\.venv\Scripts\activate
# python manage.py runserver
# python manage.py makemigrations
# python manage.py migrate